size_t/ptrdiff_t transition part 2: SFIO, hash lib, fmt*(), strmatch() (#941)
This is the second of the thickfold patch series, which enables
ksh93 to operate within a 64-bit address space. This part of the
patch series includes everything affected by changes to <ast.h>.

The parts of ksh93 affected by this commit are:
- SFIO and the associated stdio wrapper. The sfvprintf and sfvscanf
  functions are notable for bearing the brunt of the substantial
  code changes.
  - Nearly all of the code is de novo, though towards the latter
    end of development I used graphviz commits for cross-reference:
    - https://gitlab.com/graphviz/graphviz/-/commit/6451a669
    - https://gitlab.com/graphviz/graphviz/-/commit/153a8f87
    - Improved multibyte handling in sfvscanf via use of
      unsigned char*; ported from graphviz:
      https://gitlab.com/graphviz/graphviz/-/commit/cb9b35d1
      (I'm aware this function is unused, but after some manual
      testing it works well AFAICT. A bit odd that ksh93 never
      uses it once.)
  - sfseek(): Removed a wasteful double assignment for f->iosz.
    This error was introduced in 2003-06-21 ksh93o+, and likely
    would have gone unnoticed if not for the thickfold project.
    (In practice this line was virtually always optimized out by
    the compiler, so it was dead code.)

- 64-bit modernization for the AST hash library.
- Transitioned pathgetlink() to return ssize_t.
- Updated the code in print.c for compatibility with the SFIO
  64-bit modernization.
- Transitioned the strgrpmatch() family to use regflags_t to
  appease some -Wsign-conversion warnings (the type is now located
  in ast.h).
  - The now unused STR_INT flag backported from ksh93v- has been
    removed to avoid bitrot.
- Removed the unused nonstandard ssizeof() macro from ast.h.
- Made the snprintf and vsnprintf wrappers standards compliant with
  the C standard.

Change in the number of warnings on Linux when compiling with clang
using -Wsign-compare -Wshorten-64-to-32 -Wsign-conversion
-Wimplicit-int-conversion: 3,849 => 3,178 => 61 (progression from
part 1 => part 2 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/lib/libast/hash/hashlib.h
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2023 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 /*
@@ -54,10 +55,10 @@ typedef struct				/* root local pointers		*/
 
 #define _HASH_LAST_PRIVATE_ \
 	const char*	name;		/* last lookup name		*/ \
-	unsigned int	hash;		/* last lookup hash		*/
+	unsigned long	hash;		/* last lookup hash		*/
 
 #define _HASH_ROOT_PRIVATE_ \
-	int		namesize;	/* fixed name size: 0 => string	*/ \
+	size_t		namesize;	/* fixed name size: 0 => string	*/ \
 	int		meanchain;	/* resize mean chain length	*/ \
 	Hash_local_t*	local;		/* root local pointers		*/ \
 	Hash_root_t*	next;		/* next in list	of all roots	*/ \
@@ -74,7 +75,7 @@ typedef struct				/* root local pointers		*/
 #define HASHMINSIZE	(1<<4)		/* min table slots (power of 2)	*/
 #define HASHMEANCHAIN	2		/* def resize mean chain len	*/
 
-#define HASHMOD(t,h)	(h &= (t->size - 1))
+#define HASHMOD(t,h)	(h &= (unsigned long)(t->size - 1))
 #define HASHVAL(x)	((x)&~HASH_FLAGS)
 
 #define HASH(r,n,h)	if (r->local->hash) h = r->namesize ? (*r->local->hash)(n, r->namesize) : (*r->local->hash)(n);\
@@ -85,9 +86,9 @@ typedef struct				/* root local pointers		*/
 				if (r->namesize)\
 				{\
 					const char*	_hash_s2 = _hash_s1 + r->namesize;\
-					while (_hash_s1 < _hash_s2) HASHPART(h, *_hash_s1++);\
+					while (_hash_s1 < _hash_s2) HASHPART(h, (size_t)*_hash_s1++);\
 				}\
-				else while (*_hash_s1) HASHPART(h, *_hash_s1++);\
+				else while (*_hash_s1) HASHPART(h, (size_t)*_hash_s1++);\
 			}
 
 typedef struct				/* library private info		*/
