+++ src/lib/libast/stdio/sprintf.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2022 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 
@@ -26,7 +27,7 @@ sprintf(char* s, const char* fmt, ...)
 	int	v;
 
 	va_start(args, fmt);
-	v = s ? sfvsprintf(s, INT_MAX, fmt, args) : -1;
+	v = s ? (int)sfvsprintf(s, INT_MAX, fmt, args) : -1;
 	va_end(args);
 	return v;
 }
